azure_function/requirements.txt
azure-functions
azure-core
python-dotenv
pandas
snowflake-connector-python
azure-storage-blob